refactor(router): move routePath stuff in its own subfolder
