+++ modules/angular2/src/router/rules/rules.ts
@@ -6,8 +6,8 @@ import {Map} from 'angular2/src/facade/collection';
 import {RouteHandler} from './route_handlers/route_handler';
 import {Url, serializeParams} from '../url_parser';
 import {ComponentInstruction} from '../instruction';
-import {ParamRoutePath} from './param_route_path';
-import {GeneratedUrl, UrlParams} from './route_path';
+import {ParamRoutePath} from './route_paths/param_route_path';
+import {GeneratedUrl, UrlParams} from './route_paths/route_path';
 
 
 // RouteMatch objects hold information about a match between a rule and a URL
